refactor(sqlspec): give each queue store adapter one flat module
Every adapter was a package whose __init__.py did nothing but re-export a
single store.py, and stores/__init__.py re-exported all 23 classes a third
time. Each adapter is now one module named after it, and the package exports
only the base class and the factory, so every store class has exactly one
import path.

Selecting one adapter still never requires another adapter's driver: no store
module imports a driver package at module level.
