+++ src/litestar_queues/backends/sqlspec/reservation.py
@@ -159,7 +159,7 @@ class SpannerQueueReservationStore(SQLSpecTaskReservationStore):
 
     def create_schema_for_config(self, config: "Any") -> "None":
         """Create the Spanner reservation table through the native DDL operation API."""
-        from litestar_queues.backends.sqlspec.stores.spanner.store import _execute_spanner_ddl
+        from litestar_queues.backends.sqlspec.stores.spanner import _execute_spanner_ddl
 
         if not self._manage_schema:
             return
