+++ src/litestar_queues/backends/sqlspec/stores/__init__.py
@@ -1,55 +1,11 @@
-"""SQLSpec queue store implementations."""
+"""SQLSpec queue store selection.
 
-from litestar_queues.backends.sqlspec.stores.adbc import AdbcSqliteQueueStore
-from litestar_queues.backends.sqlspec.stores.aiomysql import AiomysqlQueueStore
-from litestar_queues.backends.sqlspec.stores.aiosqlite import AiosqliteQueueStore
-from litestar_queues.backends.sqlspec.stores.arrow_odbc import ArrowOdbcQueueStore
-from litestar_queues.backends.sqlspec.stores.asyncmy import AsyncmyQueueStore
-from litestar_queues.backends.sqlspec.stores.asyncpg import AsyncpgQueueStore
+Store classes are imported from the module that defines them
+(:mod:`.adapters` for family-only drivers, or the adapter's own module).
+This package exports only the base class and the factory.
+"""
+
 from litestar_queues.backends.sqlspec.stores.base import SQLSpecQueueStore
-from litestar_queues.backends.sqlspec.stores.cockroach_asyncpg import CockroachAsyncpgQueueStore
-from litestar_queues.backends.sqlspec.stores.cockroach_psycopg import (
-    CockroachPsycopgAsyncQueueStore,
-    CockroachPsycopgSyncQueueStore,
-)
-from litestar_queues.backends.sqlspec.stores.duckdb import DuckDBQueueStore
 from litestar_queues.backends.sqlspec.stores.factory import create_queue_store
-from litestar_queues.backends.sqlspec.stores.mssql_python import MssqlPythonQueueStore
-from litestar_queues.backends.sqlspec.stores.mysqlconnector import (
-    MysqlConnectorAsyncQueueStore,
-    MysqlConnectorSyncQueueStore,
-)
-from litestar_queues.backends.sqlspec.stores.oracledb import OracledbAsyncQueueStore, OracledbSyncQueueStore
-from litestar_queues.backends.sqlspec.stores.psqlpy import PsqlpyQueueStore
-from litestar_queues.backends.sqlspec.stores.psycopg import PsycopgAsyncQueueStore, PsycopgSyncQueueStore
-from litestar_queues.backends.sqlspec.stores.pymssql import PymssqlQueueStore
-from litestar_queues.backends.sqlspec.stores.pymysql import PymysqlQueueStore
-from litestar_queues.backends.sqlspec.stores.spanner import SpannerQueueStore
-from litestar_queues.backends.sqlspec.stores.sqlite import SqliteQueueStore
 
-__all__ = (
-    "AdbcSqliteQueueStore",
-    "AiomysqlQueueStore",
-    "AiosqliteQueueStore",
-    "ArrowOdbcQueueStore",
-    "AsyncmyQueueStore",
-    "AsyncpgQueueStore",
-    "CockroachAsyncpgQueueStore",
-    "CockroachPsycopgAsyncQueueStore",
-    "CockroachPsycopgSyncQueueStore",
-    "DuckDBQueueStore",
-    "MssqlPythonQueueStore",
-    "MysqlConnectorAsyncQueueStore",
-    "MysqlConnectorSyncQueueStore",
-    "OracledbAsyncQueueStore",
-    "OracledbSyncQueueStore",
-    "PsqlpyQueueStore",
-    "PsycopgAsyncQueueStore",
-    "PsycopgSyncQueueStore",
-    "PymssqlQueueStore",
-    "PymysqlQueueStore",
-    "SQLSpecQueueStore",
-    "SpannerQueueStore",
-    "SqliteQueueStore",
-    "create_queue_store",
-)
+__all__ = ("SQLSpecQueueStore", "create_queue_store")
