--- src/litestar_queues/backends/sqlspec/stores/adbc/__init__.py
@@ -1,5 +1,0 @@
-"""ADBC SQLSpec queue store."""
-
-from litestar_queues.backends.sqlspec.stores.adbc.store import AdbcSqliteQueueStore
-
-__all__ = ("AdbcSqliteQueueStore",)
