--- src/litestar_queues/backends/sqlspec/stores/aiomysql/__init__.py
@@ -1,5 +1,0 @@
-"""aiomysql SQLSpec queue store."""
-
-from litestar_queues.backends.sqlspec.stores.aiomysql.store import AiomysqlQueueStore
-
-__all__ = ("AiomysqlQueueStore",)
