--- src/litestar_queues/backends/sqlspec/stores/aiosqlite/__init__.py
@@ -1,5 +1,0 @@
-"""aiosqlite SQLSpec queue store."""
-
-from litestar_queues.backends.sqlspec.stores.aiosqlite.store import AiosqliteQueueStore
-
-__all__ = ("AiosqliteQueueStore",)
