--- src/litestar_queues/backends/sqlspec/stores/arrow_odbc/__init__.py
@@ -1,5 +1,0 @@
-"""arrow_odbc SQLSpec queue store."""
-
-from litestar_queues.backends.sqlspec.stores.arrow_odbc.store import ArrowOdbcQueueStore
-
-__all__ = ("ArrowOdbcQueueStore",)
