--- src/litestar_queues/backends/sqlspec/stores/asyncmy/__init__.py
@@ -1,5 +1,0 @@
-"""asyncmy SQLSpec queue store."""
-
-from litestar_queues.backends.sqlspec.stores.asyncmy.store import AsyncmyQueueStore
-
-__all__ = ("AsyncmyQueueStore",)
