--- src/litestar_queues/backends/sqlspec/stores/asyncpg/__init__.py
@@ -1,5 +1,0 @@
-"""asyncpg SQLSpec queue store."""
-
-from litestar_queues.backends.sqlspec.stores.asyncpg.store import AsyncpgQueueStore
-
-__all__ = ("AsyncpgQueueStore",)
