--- src/litestar_queues/backends/sqlspec/stores/cockroach_psycopg/__init__.py
@@ -1,8 +1,0 @@
-"""cockroach_psycopg SQLSpec queue stores."""
-
-from litestar_queues.backends.sqlspec.stores.cockroach_psycopg.store import (
-    CockroachPsycopgAsyncQueueStore,
-    CockroachPsycopgSyncQueueStore,
-)
-
-__all__ = ("CockroachPsycopgAsyncQueueStore", "CockroachPsycopgSyncQueueStore")
