--- src/litestar_queues/backends/sqlspec/stores/duckdb/__init__.py
@@ -1,5 +1,0 @@
-"""duckdb SQLSpec queue store."""
-
-from litestar_queues.backends.sqlspec.stores.duckdb.store import DuckDBQueueStore
-
-__all__ = ("DuckDBQueueStore",)
