--- src/litestar_queues/backends/sqlspec/stores/mssql_python/__init__.py
@@ -1,5 +1,0 @@
-"""mssql-python SQLSpec queue stores."""
-
-from litestar_queues.backends.sqlspec.stores.mssql_python.store import MssqlPythonQueueStore
-
-__all__ = ("MssqlPythonQueueStore",)
