--- src/litestar_queues/backends/sqlspec/stores/mysqlconnector/__init__.py
@@ -1,8 +1,0 @@
-"""mysqlconnector SQLSpec queue store."""
-
-from litestar_queues.backends.sqlspec.stores.mysqlconnector.store import (
-    MysqlConnectorAsyncQueueStore,
-    MysqlConnectorSyncQueueStore,
-)
-
-__all__ = ("MysqlConnectorAsyncQueueStore", "MysqlConnectorSyncQueueStore")
