--- src/litestar_queues/backends/sqlspec/stores/psqlpy/__init__.py
@@ -1,5 +1,0 @@
-"""psqlpy SQLSpec queue store."""
-
-from litestar_queues.backends.sqlspec.stores.psqlpy.store import PsqlpyQueueStore
-
-__all__ = ("PsqlpyQueueStore",)
