--- src/litestar_queues/backends/sqlspec/stores/psycopg/__init__.py
@@ -1,5 +1,0 @@
-"""psycopg SQLSpec queue store."""
-
-from litestar_queues.backends.sqlspec.stores.psycopg.store import PsycopgAsyncQueueStore, PsycopgSyncQueueStore
-
-__all__ = ("PsycopgAsyncQueueStore", "PsycopgSyncQueueStore")
