--- src/litestar_queues/backends/sqlspec/stores/pymssql/__init__.py
@@ -1,5 +1,0 @@
-"""pymssql SQLSpec queue stores."""
-
-from litestar_queues.backends.sqlspec.stores.pymssql.store import PymssqlQueueStore
-
-__all__ = ("PymssqlQueueStore",)
