--- src/litestar_queues/backends/sqlspec/stores/pymysql/__init__.py
@@ -1,5 +1,0 @@
-"""pymysql SQLSpec queue store."""
-
-from litestar_queues.backends.sqlspec.stores.pymysql.store import PymysqlQueueStore
-
-__all__ = ("PymysqlQueueStore",)
