--- src/litestar_queues/backends/sqlspec/stores/spanner/__init__.py
@@ -1,5 +1,0 @@
-"""spanner SQLSpec queue store."""
-
-from litestar_queues.backends.sqlspec.stores.spanner.store import SpannerQueueStore
-
-__all__ = ("SpannerQueueStore",)
