--- src/litestar_queues/backends/sqlspec/stores/sqlite/__init__.py
@@ -1,5 +1,0 @@
-"""sqlite SQLSpec queue store."""
-
-from litestar_queues.backends.sqlspec.stores.sqlite.store import SqliteQueueStore
-
-__all__ = ("SqliteQueueStore",)
